refactor: moved the request to the azure API to the azureAPI file
